+++ views/default/elements/forms/validation.js
@@ -39,6 +39,7 @@ define(function (require) {
 		} else {
 			message = this.catalog['_'][constraint.name];
 		}
+
 		message = message || this.catalog['_'].defaultMessage;
 		return message ? i18n.echo(message, constraint.requirementList) : this.catalog.en.defaultMessage;
 	}.bind(window.Parsley._validatorRegistry);
@@ -55,12 +56,14 @@ define(function (require) {
 		if (!$row.length) {
 			$row = this.$element.parent();
 		}
+
 		$row.addClass('elgg-field-has-errors').removeClass('elgg-field-no-errors');
 		var $errors = $row.find('.elgg-field-feedback');
 		if (!$errors.length) {
 			$errors = $('<ul class="elgg-field-feedback" />');
 			$row.append($errors);
 		}
+
 		$errors.html('');
 		$.each(errors, function (index, value) {
 			$errors.append($('<li class="elgg-field-error" />').text(value));
@@ -72,6 +75,7 @@ define(function (require) {
 		if (!$row.length) {
 			$row = this.$element.parent();
 		}
+
 		$row.removeClass('elgg-field-has-errors').addClass('elgg-field-no-errors');
 		$row.find('.elgg-field-feedback').html('');
 	});